trade_erp_real_finalfinal/trade_erp_v17_patched/trade_erp_merged/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
plotly>=5.18.0
Pillow>=10.0.0
python-docx>=1.0.0
